# docs/openapi/openapi-v2.yaml
openapi: 3.1.0
info:
  title: CircleHaven — Trust Circle API v2 (E2EE X25519/Ed25519)
  version: "2.0.0"
  description: Les corps Auth sont limités à 16 Kio et les corps Messaging à 256 Kio. Les objets d'entrée refusent les propriétés inconnues.
servers:
  - url: https://trust-circle.kavalek.fr
security:
  - bearerAuth: []
    deviceIdHeader: []
    deviceKeyVersionHeader: []
    deviceProofHeader: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token EdDSA/Ed25519 uniquement, audience trust-circle-api et typ=access.
    refreshAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Refresh token uniquement, audience trust-circle-auth et typ=refresh.
    deviceIdHeader:
      type: apiKey
      in: header
      name: X-CircleHaven-Device-Id
      description: UUID de l'appareil du compte qui produit la preuve d'accès.
    deviceKeyVersionHeader:
      type: apiKey
      in: header
      name: X-CircleHaven-Device-Key-Version
      description: Version entière positive de la clé d'identité Ed25519 de l'appareil.
    deviceProofHeader:
      type: apiKey
      in: header
      name: X-CircleHaven-Device-Proof
      description: Signature Ed25519 Base64 canonique de la transcription d'accès de 89 octets liée au jti de l'access token. Les trois en-têtes appareil sont obligatoires ensemble sur les routes Messaging, sauf enrôlement initial.
  schemas:
...
    DeviceApprovalChallenge:
      type: object
      required: [challengeId, approverDeviceId, targetDeviceId, decision, transcript, expiresAt, algorithm, target]
      properties:
        challengeId: { type: string, format: uuid }
        approverDeviceId: { type: string, format: uuid }
        targetDeviceId: { type: string, format: uuid }
        decision: { type: string, enum: [approve, reject, revoke] }
        transcript: { type: string, description: Base64 des 216 octets exacts à signer }
        expiresAt: { type: string, format: date-time }
        algorithm: { type: string, enum: [Ed25519] }
        target:
          type: object
          required: [deviceId, identityPublicKey, identityKeyVersion, platform, deviceName]
          properties:
            deviceId: { type: string, format: uuid }
            identityPublicKey: { type: string, description: Base64 canonique de la clé Ed25519 brute de 32 octets }
            identityKeyVersion: { type: integer, minimum: 1 }
            platform: { type: string, enum: [android, ios, windows, macos, unknown] }
            deviceName: { type: string }
...
paths:
...
  /api/devices/{targetDeviceId}/approvals/challenge:
    post:
      tags: [devices]
      summary: Créer une approbation, un refus ou une révocation globale à signer
      parameters:
        - in: path
          name: targetDeviceId
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [approverDeviceId, decision]
              properties:
                approverDeviceId: { type: string, format: uuid }
                decision: { type: string, enum: [approve, reject, revoke] }
      responses:
        "201":
          description: Transcription créée pour 5 minutes
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DeviceApprovalChallenge' }
        "403": { description: L'appareil approbateur n'est pas actif }
        "404": { description: Cible absente du compte courant }
        "409": { description: La cible n'est plus pending pour approve/reject ou plus active pour revoke }
        "429": { description: Limite de challenges atteinte }
  /api/devices/approvals/{challengeId}/decision:
    post:
      tags: [devices]
      summary: Appliquer une décision signée par l'appareil actif
      parameters:
        - in: path
          name: challengeId
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [signature]
              properties:
                signature: { type: string, minLength: 88, maxLength: 88, description: Base64 canonique de 64 octets Ed25519 }
      responses:
        "200":
          description: Décision appliquée atomiquement
          content:
            application/json:
              schema:
                type: object
                required: [targetDeviceId, approverDeviceId, decision, status]
                properties:
                  targetDeviceId: { type: string, format: uuid }
                  approverDeviceId: { type: string, format: uuid }
                  decision: { type: string, enum: [approve, reject, revoke] }
                  status: { type: string, enum: [active, revoked] }
        "400": { description: Encodage de signature non canonique }
        "403": { description: Signature invalide ou approbateur inactif }
        "404": { description: Challenge absent pour le compte courant }
        "409": { description: Challenge consommé ou identité/état modifié }
        "410": { description: Challenge expiré }
...